requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
openai==1.50.0
python-multipart==0.0.12
python-dotenv==1.0.1
httpx==0.27.2
